+++ noodles-fastq/src/fai.rs
@@ -2,12 +2,14 @@
 
 pub mod io;
 mod record;
-mod writer;
 
 #[deprecated(since = "0.15.0", note = "Use `fai::io::Reader` instead.")]
 pub use self::io::Reader;
 
-pub use self::{record::Record, writer::Writer};
+#[deprecated(since = "0.15.0", note = "Use `fai::io::Writer` instead.")]
+pub use self::io::Writer;
+
+pub use self::record::Record;
 
 /// A FASTQ index.
 pub type Index = Vec<Record>;